// server/core/validation/schemas.ts
export const createServerSchema = z.object({
  name: z.string().min(3).max(100),
  host: z.string().min(1).max(255),
  port: z.number().int().min(1).max(65535),
  description: z.string().max(2000).optional(),
  tags: z.array(z.string().max(50)).max(10).optional(),
  category: z.string().max(50).optional(),
  version: z.string().max(50).optional(),
  website: z.string().url().optional().or(z.literal('')),
  discord: z.string().url().optional().or(z.literal('')),
});